fix: SQLite decimal crash, untrack app.db, Railway DB diagnostics
- Fixed SQLite decimal ORDER BY crash in ExportController by casting to double
- Added .dockerignore to exclude local build artifacts and databases from Docker context
- Added *.db to .gitignore and removed old app.db from version control
- Updated Program.cs to log which database is being used (PostgreSQL vs SQLite)
- Added fatal error when Railway environment is detected but DATABASE_URL is missing
- This prevents silent SQLite fallback and makes misconfiguration obvious
- Build verified: 0 warnings, 0 errors

+++ RescateAcademico/Controllers/ExportController.cs
@@ -64,7 +64,7 @@ namespace RescateAcademico.Controllers
         {
             var alumnos = await _context.Alumnos
                 .Where(a => a.RiesgoAcademico == "Rojo" || a.RiesgoAcademico == "Amarillo")
-                .OrderByDescending(a => a.PromedioGlobal)
+                .OrderByDescending(a => (double)a.PromedioGlobal)
                 .ToListAsync();
 
             var csv = new StringBuilder();
@@ -85,7 +85,7 @@ namespace RescateAcademico.Controllers
             var alumnos = await _context.Alumnos
                 .Where(a => a.RiesgoAcademico == "Rojo" || a.RiesgoAcademico == "Amarillo")
                 .OrderBy(a => a.RiesgoAcademico)
-                .ThenByDescending(a => a.PromedioGlobal)
+                .ThenByDescending(a => (double)a.PromedioGlobal)
                 .ToListAsync();
 
             var totalAlumnos = await _context.Alumnos.CountAsync();
